nodes-chart only re-rendered on prop updates!!!
