+++ client/app/scripts/charts/node-container.js
@@ -13,6 +13,8 @@ class NodeContainer extends React.Component {
     const scaleFactor = focused ? (1 / zoomScale) : 1;
     const other = _.omit(this.props, 'dx', 'dy');
 
+    console.log('nodecontainer.render');
+
     return (
       <Motion style={{
         x: spring(dx, animConfig),
