+++ client/app/scripts/components/nodes.js
@@ -8,29 +8,10 @@ import { DelayedShow } from '../utils/delayed-show';
 import { Loading, getNodeType } from './loading';
 import { isTopologyEmpty } from '../utils/topology-utils';
 import { CANVAS_MARGINS } from '../constants/styles';
-import { nodesSelector } from '../selectors/chartSelectors';
 
 const navbarHeight = 194;
 const marginTop = 0;
 
-
-/**
- * dynamic coords precision based on topology size
- */
-function getLayoutPrecision(nodesCount) {
-  let precision;
-  if (nodesCount >= 50) {
-    precision = 0;
-  } else if (nodesCount > 20) {
-    precision = 1;
-  } else if (nodesCount > 10) {
-    precision = 2;
-  } else {
-    precision = 3;
-  }
-
-  return precision;
-}
 
 class Nodes extends React.Component {
   constructor(props, context) {
@@ -68,9 +49,8 @@ class Nodes extends React.Component {
   }
 
   render() {
-    const { nodes, topologyEmpty, gridMode, topologiesLoaded, nodesLoaded, topologies,
+    const { topologyEmpty, gridMode, topologiesLoaded, nodesLoaded, topologies,
       currentTopology } = this.props;
-    const layoutPrecision = getLayoutPrecision(nodes.size);
 
     console.log('nodes.render');
 
@@ -87,13 +67,10 @@ class Nodes extends React.Component {
         {gridMode ?
           <NodesGrid {...this.state}
             nodeSize="24"
-            nodes={nodes}
             margins={CANVAS_MARGINS}
           /> :
          <NodesChart {...this.state}
-           nodes={nodes}
            margins={CANVAS_MARGINS}
-           layoutPrecision={layoutPrecision}
            />}
       </div>
     );
@@ -116,7 +93,6 @@ function mapStateToProps(state) {
   return {
     currentTopology: state.get('currentTopology'),
     gridMode: state.get('gridMode'),
-    nodes: nodesSelector(state),
     nodesLoaded: state.get('nodesLoaded'),
     topologies: state.get('topologies'),
     topologiesLoaded: state.get('topologiesLoaded'),
